........
  r17676 | bugman | 2012-10-02 16:30:46 +0200 (Tue, 02 Oct 2012) | 3 lines
  
  Python 3 bug fix for the frame order analysis - another int division problem.
........

+++ maths_fns/frame_order/__init__.py
@@ -382,7 +382,7 @@ class Frame_order:
             raise RelaxError("The full_in_ref_frame argument " + repr(self.full_in_ref_frame) + " must be supplied.")
 
         # Tensor set up.
-        self.num_tensors = len(self.full_tensors) / 5
+        self.num_tensors = int(len(self.full_tensors) / 5)
         self.A_3D = zeros((self.num_tensors, 3, 3), float64)
         self.A_3D_bc = zeros((self.num_tensors, 3, 3), float64)
         self.A_5D_bc = zeros(self.num_tensors*5, float64)
